feat(stdutil):blocks

+++ src/stdutil/block.c
@@ -1,6 +1,6 @@
 #include "stdutil/block.h"
 
-blocks_t *init_blocks(uint8_t *blocks_start, const size_t total_size, const size_t block_size)
+blocks_t *init_blocks(uint8_t *blocks_start, const uint64_t total_size, const uint64_t block_size)
 {
     if (total_size < sizeof(blocks_t) + sizeof(block_t))
     {
@@ -16,7 +16,7 @@ blocks_t *init_blocks(uint8_t *blocks_start, const size_t total_size, const size
     return blocks;
 }
 
-block_t *block_ptr(blocks_t *blocks, size_t id)
+block_t *block_ptr(blocks_t *blocks, uint64_t id)
 {
     if (id >= (blocks->total_blocks))
     {
@@ -30,9 +30,9 @@ block_t *block_ptr(blocks_t *blocks, size_t id)
 
 block_t blocks_alloc(blocks_t *blocks)
 {
-    if (blocks->free_next_id == (size_t)-1)
+    if (blocks->free_next_id == (uint64_t)-1)
     {
-        size_t totalused_size = sizeof(blocks_t) + blocks->total_blocks * blocks->block_size;
+        uint64_t totalused_size = sizeof(blocks_t) + blocks->total_blocks * blocks->block_size;
         if (totalused_size + blocks->block_size > blocks->total_size)
         {
             LOG("Out of memory. %zu(totalused_size)= %zu(blocks_t)+ %zu(total_blocks)*%zu(block_size),when total_size %zu",
@@ -50,7 +50,7 @@ block_t blocks_alloc(blocks_t *blocks)
             *new_block = (block_t){
                 .id = blocks->total_blocks - 1,
                 .used = 1,
-                .free_next_id = (size_t)-1,
+                .free_next_id = (uint64_t)-1,
             };
 
             LOG("append new block %zu,blocks usage: %zu / %zu",
@@ -59,11 +59,11 @@ block_t blocks_alloc(blocks_t *blocks)
         }
     }
 
-    size_t free_id = blocks->free_next_id;
+    uint64_t free_id = blocks->free_next_id;
     block_t *free_block = block_ptr(blocks, free_id);
     free_block->used = 1;
     blocks->free_next_id = free_block->free_next_id;
-    free_block->free_next_id = (size_t)-1;
+    free_block->free_next_id = (uint64_t)-1;
     blocks->used_blocks++;
 
     LOG("Reusing free block %zu, Used blocks: %zu/%zu",
@@ -72,7 +72,7 @@ block_t blocks_alloc(blocks_t *blocks)
     return *free_block;
 }
 
-void blocks_free(blocks_t *blocks, size_t id)
+void blocks_free(blocks_t *blocks, uint64_t id)
 {
     LOG("block id %zu ->free", id);
     if (id >= blocks->total_blocks)
